requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
supabase>=1.0.3
python-dotenv>=0.19.0
vecs>=0.3.0
pydantic>=2.0.0
python-multipart>=0.0.5
langgraph>=0.0.15
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
openai>=1.0.0
azure-identity>=1.12.0
azure-core>=1.26.4